🎨 style(web): update viewport meta tag and add script for appearance preference
- Update viewport meta tag to prevent user scaling and add initial-scale attribute.
- Add script to check for user appearance preference and apply 'dark' class accordingly.

This commit updates the viewport meta tag in the HTML file to prevent user scaling and adds an initial-scale attribute for better mobile responsiveness. Additionally, a script is added to check for user appearance preference stored in local storage and apply the 'dark' class to the document element accordingly, improving the user experience in dark mode.